+++ eyes_common/applitools/common/config.py
@@ -10,6 +10,7 @@ from applitools.common.geometry import RectangleSize
 from applitools.common.match import ImageMatchSettings, MatchLevel
 from applitools.common.server import FailureReports, SessionType
 from applitools.common.utils import UTC, argument_guard
+from applitools.common.utils.compat import urlparse, urlunsplit
 from applitools.common.utils.converters import str2bool
 from applitools.common.utils.general_utils import get_env_with_prefix
 from applitools.common.utils.json_utils import JsonInclude
@@ -92,6 +93,38 @@ class BatchInfo(object):
         """
         del self.properties[:]
         return self
+
+
+class ProxySettings(object):
+    def __init__(
+        self,
+        host_or_url,  # type: Text
+        port=None,  # type: Optional[int]
+        username=None,  # type: Optional[Text]
+        password=None,  # type: Optional[Text]
+        scheme=None,  # type: Text
+    ):
+        # type: (...) -> None
+        if host_or_url.startswith("http://") or host_or_url.startswith("https://"):
+            parsed = urlparse(host_or_url)
+            self.host = parsed.hostname
+            self.port = port or parsed.port or 8888
+            self.username = username or parsed.username
+            self.password = password or parsed.password
+            self.scheme = scheme or parsed.scheme or "http"
+        else:
+            self.host = host_or_url
+            self.port = port or 8888
+            self.username = username
+            self.password = password
+            self.scheme = scheme or "http"
+
+    @property
+    def url(self):
+        password = ":{}".format(self.password) if self.password else ""
+        auth = "{}{}@".format(self.username, password) if self.username else ""
+        port = ":{}".format(self.port) if self.port else ""
+        return urlunsplit((self.scheme, auth + self.host + port, "", "", ""))
 
 
 @attr.s
@@ -183,6 +216,7 @@ class Configuration(object):
     features = attr.ib(
         metadata={JsonInclude.NON_NONE: True}, factory=set
     )  # type: Set[Feature]
+    proxy = attr.ib(default=None)  # type: ProxySettings
 
     @property
     def enable_patterns(self):
@@ -453,3 +487,9 @@ class Configuration(object):
     def is_feature_activated(self, feature):
         # type: (Feature) -> bool
         return feature in self.features
+
+    def set_proxy(self, proxy):
+        # type: (Optional[ProxySettings]) -> Self
+        argument_guard.is_a(proxy, ProxySettings)
+        self.proxy = proxy
+        return self