+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Added
+- Allow Python Proxy via ProxySettings object [Trello 1583](https://trello.com/c/qQAooLih)
+
 ## [4.20.1] - 2021-05-18
 ### Fixed
 - Capture regions inside manually switched frames [Trello 2136](https://trello.com/c/bOYeHaoz)